Replace data types of `app_oauth2_user_id`, `asbie_bizterm_id`, `ascc_bizterm_id`, `bbie_bizterm_id`, `bcc_bizterm_id`, `bdt_pri_restri_id`, `bdt_sc_pri_restri_id`, `bie_usage_rule_id`, `bie_user_ext_revision_id`, `business_term_id`, `cc_tag_id`, `cdt_ref_spec_id`, `cdt_sc_ref_spec_id`, `comment_id`, `dt_usage_rule_id`, `exception_id`, `message_id`, `oauth2_app_id`, `oauth2_app_scope_id`, `ref_spec_id`, `usage_rule_id`, `usage_rule_expression_id` with UUID char(36)

+++ score-http/src/main/java/org/oagi/score/gateway/http/api/cc_management/repository/CcNodeRepository.java
@@ -738,7 +738,7 @@ public class CcNodeRepository {
 
         for (BdtPriRestriRecord bdtPriRestriRecord : bdtPriRestriRecords) {
             CcBdtPriRestri ccBdtPriRestri = new CcBdtPriRestri();
-            ccBdtPriRestri.setBdtPriRestriId(bdtPriRestriRecord.getBdtPriRestriId().toBigInteger());
+            ccBdtPriRestri.setBdtPriRestriId(bdtPriRestriRecord.getBdtPriRestriId());
             if (bdtPriRestriRecord.getCdtAwdPriXpsTypeMapId() != null) {
                 ccBdtPriRestri.setType(PrimitiveRestriType.Primitive);
 
@@ -873,7 +873,7 @@ public class CcNodeRepository {
 
         for (BdtScPriRestriRecord bdtScPriRestriRecord : bdtScPriRestriRecords) {
             CcBdtScPriRestri ccBdtScPriRestri = new CcBdtScPriRestri();
-            ccBdtScPriRestri.setBdtScPriRestriId(bdtScPriRestriRecord.getBdtScPriRestriId().toBigInteger());
+            ccBdtScPriRestri.setBdtScPriRestriId(bdtScPriRestriRecord.getBdtScPriRestriId());
             if (bdtScPriRestriRecord.getCdtScAwdPriXpsTypeMapId() != null) {
                 ccBdtScPriRestri.setType(PrimitiveRestriType.Primitive);
